moved app.js to home.js, and made app.js a routing page
