+++ package.json
@@ -14,7 +14,7 @@
     "pg": "^8.11.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.1",
+    "react-router-dom": "^6.22.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
