+++ src/components/Login.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Login = () => {
+  return (
+    <>
+    <div style={{ width: '200px', height: '200px', backgroundColor: 'lightblue', border: '1px solid black' }}>
+      This is a div box.
+    </div>
+    </>
+  );
+}
+
+export default Login;
