+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-/// comment
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
